+++ plugins/main/public/components/endpoints-summary/services/get-outdated-agents.ts
@@ -1,18 +1,28 @@
 import { WzRequest } from '../../../react-services/wz-request';
 
-export const getOutdatedAgents = async ({
+export const getOutdatedAgentsService = async ({
   agentIds,
   limit,
+  q,
 }: {
   agentIds?: string[];
   limit?: number;
+  q?: string;
 }) => {
+  const qArray = q?.split(';') ?? [];
+  const qAgents = agentIds?.length
+    ? [agentIds.map(agentId => `id=${agentId}`).join(',')]
+    : [];
+  const finalQ = [...qArray, ...qAgents].join(';');
+
   const {
     data: { data },
   } = await WzRequest.apiReq('GET', '/agents/outdated', {
     params: {
-      ...(agentIds?.length
-        ? { q: `(${agentIds.map(agentId => `id=${agentId}`).join(',')})` }
+      ...(agentIds?.length || q
+        ? {
+            q: finalQ,
+          }
         : {}),
       limit,
     },